// frontend/src/enums/filters.enum.ts
export enum FilterTab {
  SORT_BY = 'SORT_BY',
  COVER = 'COVER',
  ROOM_RENT_TYPE = 'ROOM_RENT_TYPE',
  POLICY_BENEFITS = 'POLICY_BENEFITS',
  PREMIUM_PER_MONTH = 'PREMIUM_PER_MONTH',
  PORTABILITY = 'PORTABILITY',
  MATERNITY_WAIT = 'MATERNITY_WAIT',
  EXISTING_DISEASE_WAIT = 'EXISTING_DISEASE_WAIT',
  POLICY_PERIOD = 'POLICY_PERIOD',
  INSURER = 'INSURER',
  MATERNITY_COVER = 'MATERNITY_COVER',
}

export enum SortByOption {
  RELEVANCE = 'RELEVANCE',
  PREMIUM_LOW_TO_HIGH = 'PREMIUM_LOW_TO_HIGH',
  CASHLESS_HOSPITALS = 'CASHLESS_HOSPITALS',
}

export enum CoverOption {
  RECOMMENDED = 'RECOMMENDED',
  BELOW_5_LAKH = 'BELOW_5_LAKH',
  FIVE_TO_NINE_LAKH = 'FIVE_TO_NINE_LAKH',
  TEN_TO_TWENTY_FOUR_LAKH = 'TEN_TO_TWENTY_FOUR_LAKH',
  TWENTY_FIVE_TO_NINETY_NINE_LAKH = 'TWENTY_FIVE_TO_NINETY_NINE_LAKH',
  ONE_TO_TWO_CR = 'ONE_TO_TWO_CR',
  TWO_TO_SIX_CR = 'TWO_TO_SIX_CR',
  UNLIMITED = 'UNLIMITED',
}

export enum RoomRentOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  NO_ROOM_RENT_LIMIT = 'NO_ROOM_RENT_LIMIT',
  SINGLE_PRIVATE_ROOM = 'SINGLE_PRIVATE_ROOM',
  SHARED_ROOM = 'SHARED_ROOM',
}

export enum PolicyBenefitOption {
  DIABETES_COVERED = 'DIABETES_COVERED',
  PRE_HOSPITALIZATION_COVERED = 'PRE_HOSPITALIZATION_COVERED',
  POST_HOSPITALIZATION_COVERED = 'POST_HOSPITALIZATION_COVERED',
  DAY_CARE_TREATMENTS = 'DAY_CARE_TREATMENTS',
  NO_CLAIM_BONUS = 'NO_CLAIM_BONUS',
  RESTORATION_BENEFITS = 'RESTORATION_BENEFITS',
  FREE_HEALTH_CHECKUP = 'FREE_HEALTH_CHECKUP',
  DOCTOR_CONSULTATION_PHARMACY = 'DOCTOR_CONSULTATION_PHARMACY',
}

export enum ExistingDiseaseWaitOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  NO_WAITING_PERIOD = 'NO_WAITING_PERIOD',
  ONE_YEAR = 'ONE_YEAR',
  TWO_YEARS = 'TWO_YEARS',
  THREE_YEARS = 'THREE_YEARS',
}

export enum PremiumOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  BELOW_1K = 'BELOW_1K',
  ONE_TO_TWO_K = 'ONE_TO_TWO_K',
  TWO_TO_FOUR_K = 'TWO_TO_FOUR_K',
  ABOVE_FOURK = 'ABOVE_FOURK',
}

export enum PortabilityOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  ONLY_PORTABLE = 'ONLY_PORTABLE',
}

export enum MaternityWaitOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  TWO_YEARS = 'TWO_YEARS',
  THREE_YEARS = 'THREE_YEARS',
  FOUR_YEARS = 'FOUR_YEARS',
}

export enum PolicyPeriodOption {
  ONE_YEAR = 'ONE_YEAR',
  TWO_YEARS = 'TWO_YEARS',
  THREE_YEARS = 'THREE_YEARS',
  FOUR_YEARS = 'FOUR_YEARS',
  FIVE_YEARS = 'FIVE_YEARS',
}

export enum MaternityCoverOption {
  NO_PREFERENCE = 'NO_PREFERENCE',
  COVERED = 'COVERED',
}